Move out classes to its own files - normalize spaces
